+++ packages/frontend/src/components/LightdashVisualization/index.tsx
@@ -53,6 +53,7 @@ const LightdashVisualization: FC<LightdashVisualizationProps> = memo(
                 return (
                     <SimpleChart
                         className={className}
+                        isInDashboard={!!isDashboard}
                         $shouldExpand
                         data-testid={props['data-testid']}
                         {...props}
@@ -63,6 +64,7 @@ const LightdashVisualization: FC<LightdashVisualizationProps> = memo(
                     <SimplePieChart
                         className={className}
                         tileUuid={tileUuid}
+                        isInDashboard={!!isDashboard}
                         $shouldExpand
                         data-testid={props['data-testid']}
                         {...props}